% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{accession,compactProjectSummary-method}
\alias{accession,compactProjectSummary-method}
\title{Returns a project accession}
\usage{
\S4method{accession}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the accession
}
\description{
Returns a project accession
}
\author{
Jose A. Dianes
}
