% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{accession<-,compactProjectSummary-method}
\alias{accession<-,compactProjectSummary-method}
\title{Replaces a project accession}
\usage{
\S4method{accession}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the accession}
}
\description{
Replaces a project accession
}
\author{
Jose A. Dianes
}
