% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R, R/project.R
\name{affiliations,compactProjectSummary-method}
\alias{affiliations,compactProjectSummary-method}
\title{Returns a affiliations}
\usage{
\S4method{affiliations}{compactProjectSummary}(object)

\S4method{affiliations}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the affiliations

the affiliations
}
\description{
Returns a affiliations

Returns a affiliations
}
\author{
Jose A. Dianes

Jose A. Dianes
}
