% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R, R/project.R
\name{affiliations<-,compactProjectSummary-method}
\alias{affiliations<-,compactProjectSummary-method}
\title{Replaces the affiliations}
\usage{
\S4method{affiliations}{compactProjectSummary}(object) <- value

\S4method{affiliations}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{affiliations}
}
\description{
Replaces the affiliations

Replaces the affiliations
}
\author{
Jose A. Dianes

Tremayne Booker
}
