% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\docType{class}
\name{compactProjectSummary-class}
\alias{compactProjectSummary-class}
\title{compactProjectSummary represents a PRIDE Archive project dataset}
\description{
compactProjectSummary represents a PRIDE Archive project dataset
}
