% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{compactProjectSummary}
\alias{compactProjectSummary}
\title{Constructor for compactProjectSummary}
\usage{
compactProjectSummary(
  accession,
  project.title,
  project.description,
  publication.date,
  organisms,
  organism.parts,
  instruments,
  diseases,
  project.tags,
  submission.type,
  lab.PIs,
  submitters,
  affiliations
)
}
\arguments{
\item{accession}{The project accession - equivalent to project ID}

\item{project.title}{The title of the project}

\item{project.description}{The description of the project}

\item{publication.date}{The date when the project was made public by PRIDE}

\item{organisms}{The organisms of the project}

\item{organism.parts}{The organism.parts of the project}

\item{instruments}{The names of the instruments used in the project}

\item{diseases}{The diseases analyzed in the project}

\item{project.tags}{The tags for the project}

\item{submission.type}{The type of the submission, e.g. COMPLETE, PARTIAL or PRIDE}

\item{lab.PIs}{The principal investigators of the project}

\item{submitters}{The person/people who submitted the project}

\item{affiliations}{The groups the "lab.PIs" and "submitters" are affiliated with}
}
\description{
Constructor for compactProjectSummary
}
