% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\docType{class}
\name{compactProjectSummaryList-class}
\alias{compactProjectSummaryList-class}
\title{compactProjectSummaryList represents a PRIDE Archive project collection}
\description{
compactProjectSummaryList represents a PRIDE Archive project collection
}
