% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{create.filter}
\alias{create.filter}
\title{Creates the url for all of the specified filters}
\usage{
create.filter(value, field, url.string)
}
\arguments{
\item{value}{is the value to be filtered in the field. ex "Colon cancer" for field disease}

\item{field}{is the field to be filtered}

\item{url.string}{is the subset of the url for the filter}
}
\description{
Creates the url for all of the specified filters
}
\author{
Tremayne Booker
}
