% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{data.processing.protocol,ProjectSummary-method}
\alias{data.processing.protocol,ProjectSummary-method}
\title{Returns a project data processing protocol}
\usage{
\S4method{data.processing.protocol}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project data processing protocol
}
\description{
Returns a project data processing protocol
}
\author{
Jose A. Dianes
}
