% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{data.processing.protocol<-,ProjectSummary-method}
\alias{data.processing.protocol<-,ProjectSummary-method}
\title{Replaces a project data processing protocol}
\usage{
\S4method{data.processing.protocol}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the data processing protocol}
}
\description{
Replaces a project data processing protocol
}
\author{
Jose A. Dianes
}
