% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{diseases,compactProjectSummary-method}
\alias{diseases,compactProjectSummary-method}
\title{Returns a project disease names}
\usage{
\S4method{diseases}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project disease names
}
\description{
Returns a project disease names
}
\author{
Jose A. Dianes
}
