% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{diseases,ProjectSummary-method}
\alias{diseases,ProjectSummary-method}
\title{Returns a project diseases}
\usage{
\S4method{diseases}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project diseases
}
\description{
Returns a project diseases
}
\author{
Jose A. Dianes
}
