% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{diseases<-,compactProjectSummary-method}
\alias{diseases<-,compactProjectSummary-method}
\title{Replaces the project disease nanmes}
\usage{
\S4method{diseases}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the disease names}
}
\description{
Replaces the project disease nanmes
}
\author{
Jose A. Dianes
}
