% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{diseases<-,ProjectSummary-method}
\alias{diseases<-,ProjectSummary-method}
\title{Replaces a project diseases}
\usage{
\S4method{diseases}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the diseases}
}
\description{
Replaces a project diseases
}
\author{
Jose A. Dianes
}
