% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{download.by.fileDetail}
\alias{download.by.fileDetail}
\title{Downloads files given to it}
\usage{
download.by.fileDetail(file, file.dir)
}
\arguments{
\item{file}{is the file to be downloaded}

\item{file.dir}{is the directory the file is to be downloaded at}
}
\description{
Downloads files given to it
}
\details{
i dunno
}
\author{
Tremayne Booker
}
