% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{download.file.by.name}
\alias{download.file.by.name}
\title{Searches for a file by name and downloads it}
\usage{
download.file.by.name(file.name, file.dir)
}
\arguments{
\item{file.name}{is the name of the file}

\item{file.dir}{is the directory the file is to be downloaded at}
}
\description{
Searches for a file by name and downloads it
}
\details{
i dunno
}
\author{
Tremayne Booker
}
