% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{download.files.from.project}
\alias{download.files.from.project}
\title{Searches for project and downloads all files from it}
\usage{
download.files.from.project(project.accession, file.dir)
}
\arguments{
\item{project.accession}{is the project to be searched for}

\item{file.dir}{is the directory the file is to be downloaded at}
}
\description{
Searches for project and downloads all files from it
}
\details{
i dunno
}
\author{
Tremayne Booker
}
