% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{download.project.list}
\alias{download.project.list}
\title{Downloads all projects from given project list}
\usage{
download.project.list(project.list, file.dir)
}
\arguments{
\item{project.list}{The project list to be searched for}

\item{file.dir}{The directory the file is downloaded at}
}
\description{
Downloads all projects from given project list
}
\details{
The file directory must exist already.
This will create a folder for each project within the given directory.
}
\author{
Tremayne Booker
}
