% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.bytes,FileDetail-method}
\alias{file.bytes,FileDetail-method}
\title{Returns a file size}
\usage{
\S4method{file.bytes}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the file.bytes
}
\description{
Returns a file size
}
\author{
Jose A. Dianes
}
