% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.name<-,FileDetail-method}
\alias{file.name<-,FileDetail-method}
\title{Replaces a file name}
\usage{
\S4method{file.name}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the file.name}
}
\description{
Replaces a file name
}
\author{
Jose A. Dianes
}
