% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.size<-,FileDetail-method}
\alias{file.size<-,FileDetail-method}
\title{Replaces a file size}
\usage{
\S4method{file.size}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the file.size}
}
\description{
Replaces a file size
}
\author{
Jose A. Dianes
}
