% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{FileDetail}
\alias{FileDetail}
\title{Constructor for FileDetail}
\usage{
FileDetail(project.accession, file.name, file.type, file.size, download.link)
}
\arguments{
\item{project.accession}{project accession}

\item{file.name}{the name of the file}

\item{file.type}{the type of the file. e.g. RAW, PEAK, RESULT and etc}

\item{file.size}{the size of the file}

\item{download.link}{URL for downloading the file}
}
\description{
Constructor for FileDetail
}
