% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{FileDetailList}
\alias{FileDetailList}
\title{FileDetailList represents a list of FileDetails organized by their associated project}
\usage{
FileDetailList(ProjectSummary)
}
\arguments{
\item{ProjectSummary}{The file(s) associated project}
}
\description{
FileDetailList represents a list of FileDetails organized by their associated project
}
\author{
Tremayne Booker
}
