% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{project.accession}
\alias{project.accession}
\alias{project.accession<-}
\alias{file.name}
\alias{file.name<-}
\alias{file.type}
\alias{file.type<-}
\alias{file.source}
\alias{file.source<-}
\alias{file.bytes}
\alias{file.bytes<-}
\alias{download.link}
\alias{download.link<-}
\title{File Detail related generics}
\usage{
project.accession(object)

project.accession(object) <- value

file.name(object)

file.name(object) <- value

file.type(object)

file.type(object) <- value

file.source(object)

file.source(object) <- value

file.bytes(object)

file.bytes(object) <- value

download.link(object)

download.link(object) <- value
}
\arguments{
\item{object}{variable becomes an object}

\item{value}{value to be set into the object}
}
\description{
File Detail related generics
}
