% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{folder.upload.Pride}
\alias{folder.upload.Pride}
\title{Uploads given folder to specified directory on Pride}
\usage{
folder.upload.Pride(folder.path, target.dir)
}
\arguments{
\item{folder.path}{the pathway to the folder to be uploaded}

\item{target.dir}{the directory on Pride where the folder is destined}
}
\description{
Uploads given folder to specified directory on Pride
}
\details{
i dunno
}
\author{
Tremayne Booker
}
