% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{from.json.compactProjectSummary}
\alias{from.json.compactProjectSummary}
\title{Returns a compactProjectSummary instance from a JSON string representation, except
this is specifically for the search function as the JSON is formatted differently}
\usage{
from.json.compactProjectSummary(json.object)
}
\arguments{
\item{json.object}{The JSON object from Pride to be made into a compactProjectSummary}
}
\value{
The compactProjectSummary instance
}
\description{
Returns a compactProjectSummary instance from a JSON string representation, except
this is specifically for the search function as the JSON is formatted differently
}
\details{
TODO
}
\author{
Tremayne Booker
}
