% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{from.json.ProjectSummary}
\alias{from.json.ProjectSummary}
\title{Returns a ProjectSummary instance from a JSON string representation}
\usage{
from.json.ProjectSummary(json.object)
}
\arguments{
\item{json.object}{The JSON object from Pride to be made into a ProjectSummary}
}
\value{
The ProjectSummary instance
}
\description{
Returns a ProjectSummary instance from a JSON string representation
}
\details{
TODO
}
\author{
Tremayne Booker
}
