% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get.FileDetail.ProjectSummary.List}
\alias{get.FileDetail.ProjectSummary.List}
\title{Returns a FileDetailList for each of the projects given to it in project.list}
\usage{
get.FileDetail.ProjectSummary.List(project.list)
}
\arguments{
\item{project.accession}{the project accession}
}
\value{
The FileDetailList
}
\description{
Returns a FileDetailList for each of the projects given to it in project.list
}
\details{
TODO
}
\author{
Tremayne Booker
}
