% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get.FileDetailList}
\alias{get.FileDetailList}
\title{Returns a FileDetailList for each of the projects given to it in project.list}
\usage{
get.FileDetailList(project.list)
}
\arguments{
\item{project.list}{The list of ProjectSummary objects to be made into a FileDetailList}
}
\value{
The FileDetailList of all the projects in the project.list
}
\description{
Returns a FileDetailList for each of the projects given to it in project.list
}
\details{
TODO
}
\author{
Tremayne Booker
}
