% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{get.list.PeptideDetail}
\alias{get.list.PeptideDetail}
\title{Returns a list of peptides from Pride}
\usage{
get.list.PeptideDetail(page.size = 10)
}
\arguments{
\item{page.size}{The number of peptides from Pride put into the PeptideDetail list}
}
\value{
The list of protein details
}
\description{
Returns a list of peptides from Pride
}
\details{
i dunno
}
\author{
Tremayne Booker
}
