% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{get.list.ProteinDetail}
\alias{get.list.ProteinDetail}
\title{Returns a list of protein from Pride}
\usage{
get.list.ProteinDetail(page.size = 10)
}
\arguments{
\item{page.size}{is the number of proteins returned from Pride}
}
\value{
the list of protein details
}
\description{
Returns a list of protein from Pride
}
\details{
i dunno
}
\author{
Tremayne Booker
}
