% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{get.similar.projects}
\alias{get.similar.projects}
\title{Returns similar projects from the project with the given accession}
\usage{
get.similar.projects(accession, page.size = 10, page.number = 0)
}
\arguments{
\item{accession}{Project accession to find similar projects of}

\item{page.size}{The size of the pages returned}

\item{page.number}{The page to be viewed}
}
\value{
The list of similar projects
}
\description{
Returns similar projects from the project with the given accession
}
\details{
I dunno
}
\author{
Tremayne Booker
}
