% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{identified.ptm.strings,compactProjectSummary-method}
\alias{identified.ptm.strings,compactProjectSummary-method}
\title{Returns a project modification names}
\usage{
\S4method{identified.ptm.strings}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project modification names
}
\description{
Returns a project modification names
}
\author{
Jose A. Dianes
}
