% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{identified.ptm.strings,ProjectSummary-method}
\alias{identified.ptm.strings,ProjectSummary-method}
\title{Returns a project modification names}
\usage{
\S4method{identified.ptm.strings}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project modification names
}
\description{
Returns a project modification names
}
\author{
Jose A. Dianes
}
