% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{identified.ptm.strings<-,compactProjectSummary-method}
\alias{identified.ptm.strings<-,compactProjectSummary-method}
\title{Replaces the project PTMs}
\usage{
\S4method{identified.ptm.strings}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the PTMs}
}
\description{
Replaces the project PTMs
}
\author{
Jose A. Dianes
}
