% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{instruments,compactProjectSummary-method}
\alias{instruments,compactProjectSummary-method}
\title{Returns a project instrument names}
\usage{
\S4method{instruments}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project instrument names
}
\description{
Returns a project instrument names
}
\author{
Jose A. Dianes
}
