% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{instruments,ProjectSummary-method}
\alias{instruments,ProjectSummary-method}
\title{Returns a project instrument names}
\usage{
\S4method{instruments}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project instrument names
}
\description{
Returns a project instrument names
}
\author{
Jose A. Dianes
}
