% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{instruments<-,compactProjectSummary-method}
\alias{instruments<-,compactProjectSummary-method}
\title{Replaces the project instrument nanmes}
\usage{
\S4method{instruments}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the instrument names}
}
\description{
Replaces the project instrument nanmes
}
\author{
Jose A. Dianes
}
