% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{instruments<-,ProjectSummary-method}
\alias{instruments<-,ProjectSummary-method}
\title{Replaces the project instrument nanmes}
\usage{
\S4method{instruments}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the instrument names}
}
\description{
Replaces the project instrument nanmes
}
\author{
Jose A. Dianes
}
