% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R, R/project.R
\name{lab.PIs,compactProjectSummary-method}
\alias{lab.PIs,compactProjectSummary-method}
\title{Returns a lab PIs}
\usage{
\S4method{lab.PIs}{compactProjectSummary}(object)

\S4method{lab.PIs}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the lab PIs

the lab PIs
}
\description{
Returns a lab PIs

Returns a lab PIs
}
\author{
Tremayne Booker

Tremayne Booker
}
