% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R, R/project.R
\name{lab.PIs<-,compactProjectSummary-method}
\alias{lab.PIs<-,compactProjectSummary-method}
\title{Replaces the lab PIs}
\usage{
\S4method{lab.PIs}{compactProjectSummary}(object) <- value

\S4method{lab.PIs}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{lab PIs}
}
\description{
Replaces the lab PIs

Replaces the lab PIs
}
\author{
Tremayne Booker

Tremayne Booker
}
