% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{organism.parts,compactProjectSummary-method}
\alias{organism.parts,compactProjectSummary-method}
\title{Returns a project organism.parts}
\usage{
\S4method{organism.parts}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project organism.parts
}
\description{
Returns a project organism.parts
}
\author{
Jose A. Dianes
}
