% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{organism.parts,ProjectSummary-method}
\alias{organism.parts,ProjectSummary-method}
\title{Returns a project organism.parts}
\usage{
\S4method{organism.parts}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project organism.parts
}
\description{
Returns a project organism.parts
}
\author{
Jose A. Dianes
}
