% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{organism.parts<-,ProjectSummary-method}
\alias{organism.parts<-,ProjectSummary-method}
\title{Replaces the project organism.parts}
\usage{
\S4method{organism.parts}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the organism.parts}
}
\description{
Replaces the project organism.parts
}
\author{
Jose A. Dianes
}
