% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{organisms,compactProjectSummary-method}
\alias{organisms,compactProjectSummary-method}
\title{Returns a project organisms}
\usage{
\S4method{organisms}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project organisms
}
\description{
Returns a project organisms
}
\author{
Jose A. Dianes
}
