% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{organisms,ProjectSummary-method}
\alias{organisms,ProjectSummary-method}
\title{Returns a project organisms}
\usage{
\S4method{organisms}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project organisms
}
\description{
Returns a project organisms
}
\author{
Jose A. Dianes
}
