% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{organisms<-,compactProjectSummary-method}
\alias{organisms<-,compactProjectSummary-method}
\title{Replaces the project organisms}
\usage{
\S4method{organisms}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the organisms}
}
\description{
Replaces the project organisms
}
\author{
Jose A. Dianes
}
