% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{organisms<-,ProjectSummary-method}
\alias{organisms<-,ProjectSummary-method}
\title{Replaces the project organisms}
\usage{
\S4method{organisms}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the organisms}
}
\description{
Replaces the project organisms
}
\author{
Jose A. Dianes
}
