% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{page.number,compactProjectSummaryList-method}
\alias{page.number,compactProjectSummaryList-method}
\title{Returns the page number for a given protein details list}
\usage{
\S4method{page.number}{compactProjectSummaryList}(object)
}
\arguments{
\item{object}{a compactProjectSummaryList instance}
}
\value{
the page number
}
\description{
Returns the page number for a given protein details list
}
\author{
Jose A. Dianes
}
