% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{page.size,compactProjectSummaryList-method}
\alias{page.size,compactProjectSummaryList-method}
\title{Returns the page size for a given protein details list}
\usage{
\S4method{page.size}{compactProjectSummaryList}(object)
}
\arguments{
\item{object}{a compactProjectSummaryList instance}
}
\value{
the page size
}
\description{
Returns the page size for a given protein details list
}
\author{
Jose A. Dianes
}
