% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{print.PeptideDetail}
\alias{print.PeptideDetail}
\title{Prints PeptideDetail}
\usage{
\method{print}{PeptideDetail}(object)
}
\arguments{
\item{object}{is the PeptideDetail to be printed}
}
\description{
Prints PeptideDetail
}
\details{
i dunno
}
\author{
Tremayne Booker
}
