% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{print.ProteinDetail}
\alias{print.ProteinDetail}
\title{Prints ProteinDetail}
\usage{
\method{print}{ProteinDetail}(x, ...)
}
\arguments{
\item{x}{The PeptideDetail to be printed}

\item{...}{unused}
}
\description{
Prints ProteinDetail
}
\details{
i dunno
}
\author{
Tremayne Booker
}
