% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{printPeptideDetail}
\alias{printPeptideDetail}
\title{Prints Peptide Lists}
\usage{
printPeptideDetail(peptide.list)
}
\arguments{
\item{peptide.list}{is the peptideDetail to be printed}
}
\description{
Prints Peptide Lists
}
\details{
i dunno
}
\author{
Tremayne Booker
}
