% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{printProteinDetail}
\alias{printProteinDetail}
\title{Prints Protein Lists}
\usage{
printProteinDetail(protein.list)
}
\arguments{
\item{protein.list}{is the proteinDetail to be printed}
}
\description{
Prints Protein Lists
}
\details{
i dunno
}
\author{
Tremayne Booker
}
