% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.description,compactProjectSummary-method}
\alias{project.description,compactProjectSummary-method}
\title{Returns a project description}
\usage{
\S4method{project.description}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project description
}
\description{
Returns a project description
}
\author{
Jose A. Dianes
}
