% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.description<-,compactProjectSummary-method}
\alias{project.description<-,compactProjectSummary-method}
\title{Replaces a project description}
\usage{
\S4method{project.description}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the project description}
}
\description{
Replaces a project description
}
\author{
Jose A. Dianes
}
