% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{project.FileDetail}
\alias{project.FileDetail}
\title{Returns a list of PRIDE Archive files associated with a project}
\usage{
project.FileDetail(project.accession)
}
\arguments{
\item{project.accession}{the project accession}
}
\value{
The list of FileDetail objects
}
\description{
Returns a list of PRIDE Archive files associated with a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}
