% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.list,compactProjectSummaryList-method}
\alias{project.list,compactProjectSummaryList-method}
\title{Returns a project summary page as a list}
\usage{
\S4method{project.list}{compactProjectSummaryList}(object)
}
\arguments{
\item{object}{a compactProjectSummaryList instance}
}
\value{
the protein list
}
\description{
Returns a project summary page as a list
}
\author{
Jose A. Dianes
}
