% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project.list,ProjectSummaryList-method}
\alias{project.list,ProjectSummaryList-method}
\title{Returns a project summary page as a list}
\usage{
\S4method{project.list}{ProjectSummaryList}(object)
}
\arguments{
\item{object}{a ProjectSummaryList instance}
}
\value{
the protein list
}
\description{
Returns a project summary page as a list
}
\author{
Jose A. Dianes
}
