% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.tags,compactProjectSummary-method}
\alias{project.tags,compactProjectSummary-method}
\title{Returns a project tags}
\usage{
\S4method{project.tags}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project tags
}
\description{
Returns a project tags
}
\author{
Jose A. Dianes
}
