% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project.tags,ProjectSummary-method}
\alias{project.tags,ProjectSummary-method}
\title{Returns a project tags}
\usage{
\S4method{project.tags}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project tags
}
\description{
Returns a project tags
}
\author{
Jose A. Dianes
}
