% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.tags<-,compactProjectSummary-method}
\alias{project.tags<-,compactProjectSummary-method}
\title{Replaces the project tags}
\usage{
\S4method{project.tags}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the project tags}
}
\description{
Replaces the project tags
}
\author{
Jose A. Dianes
}
