% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project.tags<-,ProjectSummary-method}
\alias{project.tags<-,ProjectSummary-method}
\title{Replaces the project tags}
\usage{
\S4method{project.tags}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the project tags}
}
\description{
Replaces the project tags
}
\author{
Jose A. Dianes
}
