% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.title,compactProjectSummary-method}
\alias{project.title,compactProjectSummary-method}
\title{Returns a project title}
\usage{
\S4method{project.title}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project title
}
\description{
Returns a project title
}
\author{
Jose A. Dianes
}
