% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project.title,ProjectSummary-method}
\alias{project.title,ProjectSummary-method}
\title{Returns a project title}
\usage{
\S4method{project.title}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project title
}
\description{
Returns a project title
}
\author{
Jose A. Dianes
}
