% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{project.title<-,compactProjectSummary-method}
\alias{project.title<-,compactProjectSummary-method}
\title{Replaces a project title}
\usage{
\S4method{project.title}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the title}
}
\description{
Replaces a project title
}
\author{
Jose A. Dianes
}
