% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project.title<-,ProjectSummary-method}
\alias{project.title<-,ProjectSummary-method}
\title{Replaces a project title}
\usage{
\S4method{project.title}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the title}
}
\description{
Replaces a project title
}
\author{
Jose A. Dianes
}
