% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{accession}
\alias{accession}
\alias{accession<-}
\alias{project.title}
\alias{project.title<-}
\alias{project.description}
\alias{project.description<-}
\alias{sample.processing.protocol}
\alias{sample.processing.protocol<-}
\alias{data.processing.protocol}
\alias{data.processing.protocol<-}
\alias{publication.date}
\alias{publication.date<-}
\alias{organisms}
\alias{organisms<-}
\alias{organism.parts}
\alias{organism.parts<-}
\alias{diseases}
\alias{diseases<-}
\alias{identified.ptm.strings}
\alias{identified.ptm.strings<-}
\alias{instruments}
\alias{instruments<-}
\alias{quantification.methods}
\alias{quantification.methods<-}
\alias{project.tags}
\alias{project.tags<-}
\alias{submission.type}
\alias{submission.type<-}
\alias{lab.PIs}
\alias{lab.PIs<-}
\alias{submitters}
\alias{submitters<-}
\alias{affiliations}
\alias{affiliations<-}
\title{Project Summary related generics}
\usage{
accession(object)

accession(object) <- value

project.title(object)

project.title(object) <- value

project.description(object)

project.description(object) <- value

sample.processing.protocol(object)

sample.processing.protocol(object) <- value

data.processing.protocol(object)

data.processing.protocol(object) <- value

publication.date(object)

publication.date(object) <- value

organisms(object)

organisms(object) <- value

organism.parts(object)

organism.parts(object) <- value

diseases(object)

diseases(object) <- value

identified.ptm.strings(object)

identified.ptm.strings(object) <- value

instruments(object)

instruments(object) <- value

quantification.methods(object)

quantification.methods(object) <- value

project.tags(object)

project.tags(object) <- value

submission.type(object)

submission.type(object) <- value

lab.PIs(object)

lab.PIs(object) <- value

submitters(object)

submitters(object) <- value

affiliations(object)

affiliations(object) <- value
}
\arguments{
\item{object}{variable becomes an object}

\item{value}{value to be set into the object}
}
\description{
Project Summary related generics
}
