% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{ProjectSummary}
\alias{ProjectSummary}
\title{Constructor for ProjectSummary}
\usage{
ProjectSummary(
  accession,
  project.title,
  project.description,
  sample.processing.protocol,
  data.processing.protocol,
  publication.date,
  organisms,
  organism.parts,
  diseases,
  identified.ptm.strings,
  instruments,
  quantification.methods,
  project.tags,
  submission.type,
  lab.PIs,
  submitters,
  affiliations
)
}
\arguments{
\item{accession}{project accession}

\item{project.title}{the title of the project}

\item{project.description}{the description of the project}

\item{sample.processing.protocol}{the protocol that was used to gather the data}

\item{data.processing.protocol}{the protocol that was used to process the data}

\item{publication.date}{the date when the project was made public by PRIDE}

\item{organisms}{the organisms of the project}

\item{organism.parts}{the organism.parts of the project}

\item{identified.ptm.strings}{the names of the PTM for the project}

\item{instruments}{the names of the instruments used in the project}

\item{project.tags}{the tags for the project}

\item{submission.type}{the type of the submission, e.g. COMPLETE, PARTIAL or PRIDE}

\item{lab.PIs}{the principal investigators of the project}

\item{submitters}{the person/people who submitted the project}

\item{affiliations}{the groups the "lab.PIs" and "submitters" are affiliated with}
}
\description{
Constructor for ProjectSummary
}
