% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{ProteinDetail}
\alias{ProteinDetail}
\title{Returns a ProteinDetail from the corresponding json object}
\usage{
ProteinDetail(json.object)
}
\arguments{
\item{Accession}{the project accession}

\item{ProteinSequence}{the Protein Sequence}

\item{ProteinGroupMembers}{the protein Group members}
}
\description{
Returns a ProteinDetail from the corresponding json object
}
\details{
i dunno
}
\author{
Tremayne Booker
}
