% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{publication.date,compactProjectSummary-method}
\alias{publication.date,compactProjectSummary-method}
\title{Returns a project publication date}
\usage{
\S4method{publication.date}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project publication date
}
\description{
Returns a project publication date
}
\author{
Jose A. Dianes
}
