% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{quantification.methods,ProjectSummary-method}
\alias{quantification.methods,ProjectSummary-method}
\title{Returns a project quantification methods}
\usage{
\S4method{quantification.methods}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project quantification methods
}
\description{
Returns a project quantification methods
}
\author{
Jose A. Dianes
}
