% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{quantification.methods<-,ProjectSummary-method}
\alias{quantification.methods<-,ProjectSummary-method}
\title{Replaces a project quantification methods}
\usage{
\S4method{quantification.methods}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the quantification methods}
}
\description{
Replaces a project quantification methods
}
\author{
Jose A. Dianes
}
