% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{query,compactProjectSummaryList-method}
\alias{query,compactProjectSummaryList-method}
\title{Returns the query term used to retrieve the compactProjectSummaryList}
\usage{
\S4method{query}{compactProjectSummaryList}(object)
}
\arguments{
\item{object}{a compactProjectSummaryList instance}
}
\value{
the query term used to retrieve this compactProjectSummaryList
}
\description{
Returns the query term used to retrieve the compactProjectSummaryList
}
\author{
Jose A. Dianes
}
