% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{sample.processing.protocol,ProjectSummary-method}
\alias{sample.processing.protocol,ProjectSummary-method}
\title{Returns a project sample processing protocol}
\usage{
\S4method{sample.processing.protocol}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project sample processing protocol
}
\description{
Returns a project sample processing protocol
}
\author{
Jose A. Dianes
}
