% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{sample.processing.protocol<-,ProjectSummary-method}
\alias{sample.processing.protocol<-,ProjectSummary-method}
\title{Replaces a project sample processing protocol}
\usage{
\S4method{sample.processing.protocol}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the sample processing protocol}
}
\description{
Replaces a project sample processing protocol
}
\author{
Jose A. Dianes
}
