% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.file.by.name}
\alias{search.file.by.name}
\title{Finds all the files with the user given file name}
\usage{
search.file.by.name(file.name)
}
\arguments{
\item{file.name}{the name of the file user wishes to find}
}
\value{
The FileDetail requested
}
\description{
Finds all the files with the user given file name
}
\details{
I dunno
}
\author{
Tremayne Booker
}
