% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.FileDetail.loop}
\alias{search.FileDetail.loop}
\title{Extension of search.project.list if the user would like AND matches}
\usage{
search.FileDetail.loop(
  project,
  keywords,
  all,
  filetype,
  file.size.min,
  file.size.max,
  use.regex
)
}
\arguments{
\item{keywords}{the word or words the user is searching for in files}

\item{project.list}{the list of projects the user wishes to search the files through}
}
\value{
the project list of successful matches
}
\description{
Extension of search.project.list if the user would like AND matches
}
\details{
i dunno
}
\author{
Tremayne Booker
}
