% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.FileDetail}
\alias{search.FileDetail}
\title{Searches a list of projects for keyword matches}
\usage{
search.FileDetail(
  project.list,
  keywords = "",
  filetype = " ",
  file.size.min = 0,
  file.size.max = 99999999999,
  all = FALSE,
  use.regex = FALSE
)
}
\arguments{
\item{project.list}{The list of projects the user wishes to search the files through}

\item{keywords}{The word or words the user is searching for in files}

\item{filetype}{The category of file. Categories are described by Pride}

\item{file.size.min}{The lower range of file memory size}

\item{file.size.max}{The upper range of file memory size}

\item{all}{Determines whether the to match using ALL or ANY logic. Defaults to ANY}

\item{use.regex}{Determines whether the keywords are regex patterns or should be matched literally}
}
\value{
the FileDetailList list of successful matches
}
\description{
Searches a list of projects for keyword matches
}
\details{
This functionality is not built natively into the Pride web service.
Searching through a ProjectSummary list will take a long time.
For multiple searches on the same ProjectSummary list, it is recommended to create
FileDetailList first, then search through that. This can be done by inputting the
ProjectSummary list into "get.FileDetailList()".
While this function can accept ProjectSummaries, it will always return a FileDetailList.
This is to ensure recursive searches are faster.
}
\author{
Tremayne Booker
}
