% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.project.list.and}
\alias{search.project.list.and}
\title{Extension of search.project.list if the user would like AND matches}
\usage{
search.project.list.and(project.list, keywords)
}
\arguments{
\item{project.list}{the list of projects the user wishes to search the files through}

\item{keywords}{the word or words the user is searching for in files}
}
\value{
the project list of successful matches
}
\description{
Extension of search.project.list if the user would like AND matches
}
\details{
i dunno
}
\author{
Tremayne Booker
}
