% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.project.list}
\alias{search.project.list}
\title{Searches a list of projects for keyword matches}
\usage{
search.project.list(project.list, keywords, and = FALSE)
}
\arguments{
\item{project.list}{the list of projects the user wishes to search the files through}

\item{keywords}{the word or words the user is searching for in files}

\item{and}{TRUE/FALSE whether the user wants to match using AND or OR. Defaults to OR}
}
\value{
the project list of successful matches
}
\description{
Searches a list of projects for keyword matches
}
\details{
i dunno
}
\author{
Tremayne Booker
}
