% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{search.ProjectSummary.and}
\alias{search.ProjectSummary.and}
\title{Returns a series of PRIDE Archive projects
to satisfy a given query. This is actually a
query filtered version of project_list.
Multiple keywords are matched by AND logic}
\usage{
search.ProjectSummary.and(
  keywords,
  page.size = 10,
  page.number = 0,
  sort.direction = "DESC"
)
}
\arguments{
\item{keywords}{The query term or terms}

\item{page.size}{The maximum number of search results}

\item{page.number}{The page of the list}

\item{sort.direction}{the direction the list is sorted by}
}
\value{
The search results in a list of objects
}
\description{
Returns a series of PRIDE Archive projects
to satisfy a given query. This is actually a
query filtered version of project_list.
Multiple keywords are matched by AND logic
}
\details{
TODO
}
\author{
Jose A. Dianes
}
