% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{search.ProjectSummary}
\alias{search.ProjectSummary}
\title{Returns a series of PRIDE Archive projects to satisfy a given query.}
\usage{
search.ProjectSummary(
  keywords = "",
  page.size = 10,
  page.number = 0,
  sort.direction = "DESC",
  organism = "",
  organism.part = "",
  instrument = "",
  disease = "",
  modification = "",
  project.tag = "",
  project.keyword = "",
  country = ""
)
}
\arguments{
\item{keywords}{The query term or terms}

\item{page.size}{The number of projects from Pride put into the compactProjectSummary list. Limit is 100}

\item{page.number}{The number of the page retrieved from Pride}

\item{sort.direction}{The direction the list is sorted by. Can be DESC (descending) or ASC (ascending)}

\item{organism}{The filterable field organism}

\item{organism.part}{The filterable field organism.part}

\item{instrument}{The filterable field instrument}

\item{disease}{The filterable field disease}

\item{modification}{The filterable field modification. Referred to as "identified PTM strings" elsewhere in package. Note that it is not present in the compactProjectSummary}

\item{project.tag}{The filterable field project tag. Referred to as "tags" elsewhere in package}

\item{project.keyword}{The filterable field project keyword. Not the same as the parameter "keywords". This field are keywords submitters specified to include in their project}

\item{country}{The filterable field country. Note that it is not present in the compactProjectSummary}
}
\value{
The search results in a list of objects
}
\description{
Returns a series of PRIDE Archive projects to satisfy a given query.
}
\details{
"Keywords" searched for search through every piece of metadata in the project for matches. It can accept multiple words as a character vector
They are also matched using ANY logic, meaning that if a keyword is present anywhere in the project
it will return that project even if other keywords are missing.
All parameters after "sort.direction" are filters that only check their respective fields for matches.
The filters must be matched exactly how they appear in the meta data, otherwise it will not return projects correctly.
For example, if you enter "Coloncancer" instead of "Colon cancer" it will not return any projects. It is not case sensitive, though.
}
\author{
Tremayne Booker
}
