% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{show,compactProjectSummary-method}
\alias{show,compactProjectSummary-method}
\title{Show the print-out version of the content in a compactProjectSummary}
\usage{
\S4method{show}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a given compactProjectSummary}
}
\description{
Show the print-out version of the content in a compactProjectSummary
}
