% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{show,ProjectSummary-method}
\alias{show,ProjectSummary-method}
\title{Show the print-out version of the content in a ProjectSummary}
\usage{
\S4method{show}{ProjectSummary}(object)
}
\arguments{
\item{object}{a given ProjectSummary to be printed}
}
\description{
Show the print-out version of the content in a ProjectSummary
}
