% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{submission.type,compactProjectSummary-method}
\alias{submission.type,compactProjectSummary-method}
\title{Returns a project submission type}
\usage{
\S4method{submission.type}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the project submission type
}
\description{
Returns a project submission type
}
\author{
Jose A. Dianes
}
