% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R
\name{submission.type<-,compactProjectSummary-method}
\alias{submission.type<-,compactProjectSummary-method}
\title{Replaces the project submission type}
\usage{
\S4method{submission.type}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{the submission type}
}
\description{
Replaces the project submission type
}
\author{
Jose A. Dianes
}
