% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R, R/project.R
\name{submitters,compactProjectSummary-method}
\alias{submitters,compactProjectSummary-method}
\title{Returns a submitters}
\usage{
\S4method{submitters}{compactProjectSummary}(object)

\S4method{submitters}{compactProjectSummary}(object)
}
\arguments{
\item{object}{a compactProjectSummary}
}
\value{
the submitters

the submitters
}
\description{
Returns a submitters

Returns a submitters
}
\author{
Jose A. Dianes

Jose A. Dianes
}
