% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactProject.R, R/project.R
\name{submitters<-,compactProjectSummary-method}
\alias{submitters<-,compactProjectSummary-method}
\title{Replaces the submitters}
\usage{
\S4method{submitters}{compactProjectSummary}(object) <- value

\S4method{submitters}{compactProjectSummary}(object) <- value
}
\arguments{
\item{object}{a compactProjectSummary}

\item{value}{submitters}
}
\description{
Replaces the submitters

Replaces the submitters
}
\author{
Jose A. Dianes

Jose A. Dianes
}
