% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{download.files.from.project.list}
\alias{download.files.from.project.list}
\title{Downloads all projects from given project list}
\usage{
download.files.from.project.list(project.list, file.dir)
}
\arguments{
\item{file.dir}{is the directory the file is to be downloaded at}

\item{project.accession}{is the project to be searched for}
}
\description{
Downloads all projects from given project list
}
\details{
i dunno
}
\author{
Tremayne Booker
}
