% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{FileDetailList}
\alias{FileDetailList}
\title{FileDetailList represents a list of FileDetails organized by their associated project}
\usage{
FileDetailList(ProjectSummary)
}
\arguments{
\item{ProjectSummary}{is the file(s) associated project}

\item{FileDetailList}{is the list of file details that are within the project}

\item{file.count}{is the amount of files associated with the project}
}
\description{
FileDetailList represents a list of FileDetails organized by their associated project
}
\author{
Tremayne Booker
}
