% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{from.json.FileDetail}
\alias{from.json.FileDetail}
\title{Returns a FileDetail instance from a JSON string representation}
\usage{
from.json.FileDetail(json.object)
}
\arguments{
\item{json.object}{The JSON object to be made into a FileDetail object}
}
\value{
The AssayDetail instance
}
\description{
Returns a FileDetail instance from a JSON string representation
}
\details{
TODO
}
\author{
Jose A. Dianes
}
