% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get.FileDetail}
\alias{get.FileDetail}
\title{Returns a list of PRIDE Archive files associated with a project}
\usage{
get.FileDetail(project.accession)
}
\arguments{
\item{project.accession}{The project accession}
}
\value{
The list of FileDetail objects of the project
}
\description{
Returns a list of PRIDE Archive files associated with a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}
