% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{get.PeptideDetail.accession}
\alias{get.PeptideDetail.accession}
\title{Returns a list of peptides from a specific project from Pride}
\usage{
get.PeptideDetail.accession(
  accession,
  page.size = 10,
  page.number = 0,
  sort.direction = "DESC"
)
}
\arguments{
\item{accession}{is the project accession}
}
\value{
the list of protein details
}
\description{
Returns a list of peptides from a specific project from Pride
}
\details{
i dunno
}
\author{
Tremayne Booker
}
