% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{get.ProteinDetail.accession}
\alias{get.ProteinDetail.accession}
\title{Returns a list of proteins from a specific project from Pride}
\usage{
get.ProteinDetail.accession(accession)
}
\arguments{
\item{accession}{is the project accession}
}
\value{
the list of protein details
}
\description{
Returns a list of proteins from a specific project from Pride
}
\details{
i dunno
}
\author{
Tremayne Booker
}
