% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{list.ProjectSummary}
\alias{list.ProjectSummary}
\title{Returns a list of PRIDE Archive project summaries}
\usage{
list.ProjectSummary(page.size = 10, page.number = 0)
}
\arguments{
\item{count}{the maximum number of projects}
}
\value{
The list of ProjectSummary objects
}
\description{
Returns a list of PRIDE Archive project summaries
}
\details{
TODO
}
\author{
Jose A. Dianes
}
