% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide.R
\name{PeptideDetail}
\alias{PeptideDetail}
\title{Returns a ProteinDetail from the corresponding json object}
\usage{
PeptideDetail(json.object)
}
\arguments{
\item{json.object}{The JSON object from Pride to be made into a PeptideDetail}
}
\description{
Returns a ProteinDetail from the corresponding json object
}
\details{
i dunno
}
\author{
Tremayne Booker
}
