% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.FileDetail.loop}
\alias{search.FileDetail.loop}
\title{The function that determines whether the project has successfully matching files}
\usage{
search.FileDetail.loop(
  project,
  keywords = "",
  filetype = " ",
  file.size.min = 0,
  file.size.max = 99999999999,
  all = FALSE,
  use.regex = FALSE
)
}
\arguments{
\item{project}{The project to search the files through}

\item{keywords}{The word or words the user is searching for in files}

\item{filetype}{The category of file. Categories are described by Pride}

\item{file.size.min}{The lower range of file memory size}

\item{file.size.max}{The upper range of file memory size}

\item{all}{Determines whether the to match using ALL or ANY logic. Defaults to ANY}

\item{use.regex}{Determines whether the keywords are regex patterns or should be matched literally}
}
\value{
the project list of successful matches
}
\description{
The function that determines whether the project has successfully matching files
}
\details{
i dunno
}
\author{
Tremayne Booker
}
