% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{search.FileDetail}
\alias{search.FileDetail}
\title{Searches a list of projects for keyword matches}
\usage{
search.FileDetail(
  project.list,
  keywords,
  all = FALSE,
  file.size.min = 0,
  file.size.max = 99999999999
)
}
\arguments{
\item{project.list}{the list of projects the user wishes to search the files through}

\item{keywords}{the word or words the user is searching for in files}

\item{and}{TRUE/FALSE whether the user wants to match using AND or OR. Defaults to OR}

\item{file.size.low}{is the lower range of file memory size the user can specify}

\item{file.size.high}{is the upper range of file memory size the user can specify}
}
\value{
the project list of successful matches
}
\description{
Searches a list of projects for keyword matches
}
\details{
TODO
}
\author{
Tremayne Booker
}
