% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{show,FileDetail-method}
\alias{show,FileDetail-method}
\title{print out the details of the FileDetail on screen}
\usage{
\S4method{show}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail object}
}
\description{
print out the details of the FileDetail on screen
}
\author{
Jose A. Dianes
}
